+++ lib/core/include/metashell/core/engine_entry.hpp
@@ -24,6 +24,7 @@
 
 #include <boost/filesystem/path.hpp>
 
+#include <functional>
 #include <map>
 #include <string>
 #include <vector>
@@ -35,10 +36,12 @@ namespace metashell
     class engine_entry
     {
     public:
-      engine_entry(engine_factory factory_,
-                   std::string args_,
-                   data::markdown_string description_,
-                   std::vector<data::feature> features_);
+      engine_entry(
+          engine_factory factory_,
+          std::string args_,
+          data::markdown_string description_,
+          std::vector<data::feature> features_,
+          std::function<bool(const std::vector<std::string>&)> this_engine_);
 
       std::unique_ptr<iface::engine>
       build(const data::config& config_,
@@ -55,12 +58,18 @@ namespace metashell
 
       const std::vector<data::feature>& features() const;
 
+      bool usable_by_auto() const;
+      bool this_engine(const std::vector<std::string>&) const;
+
     private:
       engine_factory _factory;
       std::string _args;
       data::markdown_string _description;
       std::vector<data::feature> _features;
+      std::function<bool(const std::vector<std::string>&)> _this_engine;
     };
+
+    std::function<bool(const std::vector<std::string>&)> never_used_by_auto();
 
     std::string list_features(const engine_entry& engine_);
     data::markdown_string